+++ pkg/js/errors.go
@@ -1,3 +1,5 @@
+//go:build js
+
 package js
 
 import (
